tela.mak: Re-enable command-line assignments
tela.mak uses a 'MAKEOVERRIDES:=' statement to remove all variable
assignments specified on the make command line. This is required to
enable 'make TESTS=subdir' invocations, where a subdirectory Makefile
specifies its own set of TESTS.

As a side-effect, this prevents the use of any command-line assignment
that overrides an existing Makefile variable such as the CC variable
used to identify the C compiler to use.

Fix this by using a modified filter-out rule on MAKEOVERRIDES that only
removes the TESTS= assignment instead. The filter rule takes into
account the special encoding of space characters within the
MAKEOVERRIDES variable (' ' => '\ ').

Signed-off-by: Peter Oberparleiter <oberpar@linux.ibm.com>