doc/resources/tape: Correct variable name for char devices
The char device names for tape devices are documented incorrectly in the
YAML output as well as for the environment variables. The device names
are not child nodes of CCW.

Fix the indentation in the YAML output and fix the environment variable
names by removing the CCW portion. Move the variables and the
corresponding documentation details to correlate with the YAML output
order.

While at it, format the documentation of the variables to use a 80
character line break.

Signed-off-by: Jan Höppner <hoeppner@linux.ibm.com>
Reviewed-by: Peter Oberparleiter <oberpar@linux.ibm.com>